fix button structure
